requirements.txt
--index-url https://pypi.python.org/simple/

requests>=2.5.1
python-bitcoinrpc>=1.0
-e .
